Add files via upload
This is the configuration information contained in the .m file (MATLAB) used with Kilosort2 (https://github.com/MouseLand/Kilosort) to detect Type 1 events in the PEDOT:PSS recordings from the surface of the cortex.  Further details are included here:

We used the newest version of Kilosort, appropriately named Kilosort2 (https://github.com/MouseLand/Kilosort). The mean waveforms for these neurons are swept repeatedly through the voltage data as a template, and the newly detected spikes which match these templates are subtracted away, leaving behind the residual data which can be iterated through again. The software also considers the spatial arrangement of the electrode. As such, we included a specific channel map for each electrode used as well as a consistent configuration file between participant data when sorting with Kilosort2 for detection. These data were passed in with default Kilosort2 parameters for configuration (see PEDOTConfigDetectingType1Events.m file) except that the acceptable firing rate for a unit was pushed below 0.1Hz.